move memory.c functions into main.c
